Moved forms/Planform dir to components
